# authserver/src/main/resources/static/api-doc.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8090/auth","description":"Generated server url"}],"paths":{"/admin/user/{userId}/toggle-status":{"put":{"tags":["admin-controller"],"operationId":"toggleUserStatus","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/admin/update-user":{"put":{"tags":["admin-controller"],"operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/api/refresh":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/api/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/admin/logout-force":{"post":{"tags":["admin-controller"],"operationId":"forceLogout","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/appStatus":{"get":{"tags":["app-status-controller"],"operationId":"status","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusDTO"}}}}}}},"/api/verify-email":{"get":{"tags":["auth-controller"],"operationId":"verifyEmail","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","minLength":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/admin/users":{"get":{"tags":["admin-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/admin/user/{userId}":{"get":{"tags":["admin-controller"],"operationId":"getUserById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/admin/me":{"get":{"tags":["admin-controller"],"operationId":"getMe","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/delete":{"delete":{"tags":["auth-controller"],"operationId":"delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/admin/delete-user/{id}":{"delete":{"tags":["admin-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"role":{"type":"string","enum":["STUDENT","INSTRUCTOR","ADMIN"]},"enabled":{"type":"boolean"},"deleted":{"type":"boolean"}}},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*#?&]).{8,}$"},"role":{"type":"string","minLength":1}},"required":["email","role"]},"JwtResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"message":{"type":"string"}}},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","password"]},"LoginResponse":{"type":"object","properties":{"email":{"type":"string"},"role":{"type":"string"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"StatusDTO":{"type":"object","properties":{"status":{"type":"string"}}},"DeleteRequest":{"type":"object","properties":{"accessToken":{"type":"string","minLength":1}},"required":["accessToken"]}}}}